Implement the start of caching
Also, change the way the dictionary Hash
behaves. #default_proc removed as there
was no way to implement it effectively at
the moment for caching; it wasn't saving
anything by way of memoray usage re
dictionary size, and it was causing
stack level too deep errors the way it
was currently coded as it had to call
back into the Hash to update the cache.
I'll revisit this when I redo the caching.