Refactor WordFilterManagable
Added method #clear_filters to remove all word filters.

Added method #filters_on= that takes true/false to turn
all word filters on/off.

Change WordFilterManagable#add_filters to:

Require a block to be passed if arg word_filters is
not provided.

If a block is provided, it word_filters must be
an Array of word filter class types and not be
blank?